Add bytes-based solution and shims
