+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/stringFunctions.scala
@@ -92,89 +92,6 @@ case class GpuOctetLength(child: Expression) extends GpuUnaryExpression with Exp
     input.getBase.getByteCount
 }
 
-case class GpuAscii(child: Expression) extends GpuUnaryExpression with ImplicitCastInputTypes 
-    with NullIntolerant {
-
-  override def dataType: DataType = IntegerType
-  override def inputTypes: Seq[AbstractDataType] = Seq(StringType)
-
-  private def utf8CodePointsToAscii(codePoints: ColumnVector): ColumnVector = {
-    // Currently we only support ASCII characters, so we need to convert the UTF8 code points
-    // to ASCII code points. Results for code points outside range [0, 255] are undefined.
-    // seg A: 0 <= codePoints < 128, already ASCII
-    // seg B: 49792 <= codePoints < 49856, ASCII = codePoints - 49664
-    // seg C: 50048 <= codePoints < 50112, ASCII = codePoints - 49856
-    //
-    // To reduce cuDF API calling, following algorithm will be performed:
-    // 1. For anything above 49792, we subtract 49664, now seg A and B are correct.
-    // 2. seg C: 50048 <= current + 49664 < 50112 => 384 <= current < 448, ASCII = current - 192
-    // So for anything above 384, we subtract 192, now seg C is correct too.
-    val greaterThan49792 = withResource(Scalar.fromInt(49792)) { segBLeftEnd =>
-      codePoints.greaterOrEqualTo(segBLeftEnd)
-    }
-    val segAB = withResource(greaterThan49792) { _ =>
-      val sub1 = withResource(Scalar.fromInt(49664)) { segBValue =>
-        codePoints.sub(segBValue)
-      }
-      withResource(sub1) { _ =>
-        greaterThan49792.ifElse(sub1, codePoints)
-      }
-    }
-    withResource(segAB) { _ =>
-      val geraterThan384 = withResource(Scalar.fromInt(384)) { segCLeftEnd =>
-        segAB.greaterOrEqualTo(segCLeftEnd)
-      }
-      withResource(geraterThan384) { _ =>
-        val sub2 = withResource(Scalar.fromInt(192)) { segCValue =>
-          segAB.sub(segCValue)
-        }
-        withResource(sub2) { _ =>
-          geraterThan384.ifElse(sub2, segAB)
-        }
-      }
-    }
-  }
-
-  override def doColumnar(input: GpuColumnVector): ColumnVector = {
-    // replace empty strings with 'NUL' (which will convert to ascii 0)
-    val emptyMask = withResource(Scalar.fromString("")) { emptyScalar =>
-      input.getBase.equalTo(emptyScalar)
-    }
-    val emptyReplaced = withResource(emptyMask) { _ =>
-      // replace empty strings with 'NUL' (which will convert to ascii 0)
-      withResource(Scalar.fromString('\u0000'.toString)) { zeroScalar =>
-        emptyMask.ifElse(zeroScalar, input.getBase)
-      }
-    }
-    // replace nulls with 'n' and save the null mask
-    val nullMask = closeOnExcept(emptyReplaced) { _ =>
-      input.getBase.isNull
-    }
-    withResource(nullMask) { _ =>
-      val nullsReplaced = withResource(emptyReplaced) { _ =>
-        withResource(Scalar.fromString("n")) { nullScalar =>
-          nullMask.ifElse(nullScalar, emptyReplaced)
-        }
-      }
-      val substr = withResource(nullsReplaced) { _ =>
-        nullsReplaced.substring(0, 1)
-      }
-      val codePoints = withResource(substr) { _ =>
-        substr.codePoints()
-      }
-      val segABC = withResource(codePoints) { _ =>
-        utf8CodePointsToAscii(codePoints)
-      }
-      // replace nulls with null
-      withResource(segABC) { _ =>
-        withResource(Scalar.fromNull(DType.INT32)) { nullScalar =>
-          nullMask.ifElse(nullScalar, segABC)
-        }
-      }
-    }
-  }
-}
-
 case class GpuStringLocate(substr: Expression, col: Expression, start: Expression)
   extends GpuTernaryExpressionArgsScalarAnyScalar
       with ImplicitCastInputTypes {
